updated message images to be cached properly
